(util) Some work on porting to AVR; see full commit message for caveats!
Note that this branch currently DOES NOT COMPILE for AVR, though it works
fine otherwise.

Because there's no de-facto standard C++ library available for the AVR
architecture (and I'd prefer to avoid the overhead of a full-blown C++
standard library anyway in that case), we'll be adapting the RoboCRISP
code to that architecture using mostly minimalist types/etc.

On AVR, we have the following API differences [vs. a hosted environment]
so far:

  - Any classes derived from RefCountedObject in a hosted environment will
    not do so on the AVR;

  - All `Buffer` functions that use heap-allocated memory are unavailable;
    and

  - `PeriodicScheduleSlot` uses a simple static-capacity array type
    `StaticArray` instead of `std::forward_list` to manage its actions.
    It's likely that we'll be doing something similar for signals as well.

In general, our goals for the AVR port are

  - avoid dynamic memory allocation where feasible,

  - fall back to simple alternatives for STL containers,

  - make use of the hardware functions to replace hosted I/O code, and

  - preserve the API (i.e. keep it consistent) to the maximum extent
    possible.